pyproject.toml
[project]
name = "isitprime"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.38.0",
    "asyncio>=3.4.3",
    "matplotlib>=3.9.2",
    "supabase>=2.7.4",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
